+++ src/bin/annis.rs
@@ -49,6 +49,7 @@ impl CommandCompleter {
         known_commands.insert("find".to_string());
         known_commands.insert("frequency".to_string());
         known_commands.insert("plan".to_string());
+        known_commands.insert("use_disk".to_string());
         known_commands.insert("use_parallel".to_string());
         known_commands.insert("quirks_mode".to_string());
         known_commands.insert("info".to_string());
@@ -111,6 +112,7 @@ struct AnnisRunner {
     limit: Option<usize>,
     data_dir: PathBuf,
     use_parallel_joins: bool,
+    use_disk: bool,
     query_language: QueryLanguage,
 }
 
@@ -121,6 +123,7 @@ impl AnnisRunner {
             current_corpus: vec![],
             data_dir: PathBuf::from(data_dir),
             use_parallel_joins: true,
+            use_disk: false,
             query_language: QueryLanguage::AQL,
             offset: 0,
             limit: None,
@@ -193,6 +196,7 @@ impl AnnisRunner {
                 "find" => self.find(&args),
                 "frequency" => self.frequency(&args),
                 "use_parallel" => self.use_parallel(&args),
+                "use_disk" => self.use_disk(&args),
                 "quirks_mode" => self.quirks_mode(&args),
                 "info" => self.info(),
                 "quit" | "exit" => return false,
@@ -232,6 +236,7 @@ impl AnnisRunner {
                 &PathBuf::from(path),
                 ImportFormat::RelANNIS,
                 overwritten_corpus_name,
+                self.use_disk,
             )?;
         let load_time = t_before.elapsed();
         if let Ok(t) = load_time {
@@ -528,6 +533,17 @@ impl AnnisRunner {
             println!("Join parallization is disabled");
         }
 
+        Ok(())
+    }
+
+    fn use_disk(&mut self, args: &str) -> Result<()> {
+        let new_val = match args.trim().to_lowercase().as_str() {
+            "on" | "true" => true,
+            "off" | "false" => false,
+            _ => return Err(format!("unknown argument \"{}\"", args).into()),
+        };
+
+        self.use_disk = new_val;
         Ok(())
     }