+++ src/capi/corpusstorage.rs
@@ -451,6 +451,7 @@ pub extern "C" fn annis_cs_import_from_fs(
     path: *const libc::c_char,
     format: ImportFormat,
     corpus: *const libc::c_char,
+    disk_based: bool,
     err: *mut *mut ErrorList,
 ) -> *mut libc::c_char {
     let cs: &mut CorpusStorage = cast_mut!(ptr);
@@ -463,7 +464,12 @@ pub extern "C" fn annis_cs_import_from_fs(
     let path: &str = &cstr!(path);
 
     let corpus_name: String = try_cerr!(
-        cs.import_from_fs(&PathBuf::from(path), format, override_corpus_name),
+        cs.import_from_fs(
+            &PathBuf::from(path),
+            format,
+            override_corpus_name,
+            disk_based
+        ),
         err,
         std::ptr::null_mut()
     );